pyproject.toml
[project]
name = "agent-tools"
version = "0.1.0"
description = "Self-modifying tool registry for AI agents"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "mcp>=1.0.0",
]
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [
    {name = "Michael Pruitt", email = "mpruitt@redhat.com"}
]
keywords = ["ai", "agents", "tools", "mcp", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.urls]
Homepage = "https://github.com/amp-rh/agent-tools"
Repository = "https://github.com/amp-rh/agent-tools"
Issues = "https://github.com/amp-rh/agent-tools/issues"

[project.scripts]
agent-tools = "agent_tools.cli:main"
agent-tools-server = "agent_tools.server:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "ruff>=0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_tools"]

[tool.hatch.build.targets.wheel.shared-data]
"agent-tools.yaml" = "share/agent-tools/agent-tools.yaml"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
